Refactor API VIEWS & URLS
