+++ BQE_api/glossary/urls.py
@@ -2,17 +2,18 @@ from django import urls
 from django.urls import path
 from django.urls.conf import include
 from .views import *
-from rest_framework import routers, serializers, viewsets
+from glossary.api.urls import *
 
-router = routers.DefaultRouter()
-router.register(r'word', WordViewSet)
+app_name = 'glossary'
 
 
 urlpatterns = [
 
-    path('', include(router.urls)),
-    path('data', index, name='index'),
+    path('', index, name='index'),
     path('seed_glossary', seed_glossary, name='seed_glossary'),
     path('delete_glossary', delete_glossary, name='delete_glossary'),
+
+    #REST FRAMEWORK
+    path('api/', include('glossary.api.urls','api'))
     
 ]
